requirements.txt
boto3==1.35.85
pandas==2.2.3
psycopg2==2.9.10
python-dotenv